requirements.txt
Pillow == 8.2.0
python-resize-image == 1.1.19
termcolor==1.1.0